Rename stack-manager to package-manager
This updates the manage and unpack commands to be
part of the package-manager instead of stack-manager
to more closely reflect the renaming of Stack types
to Package types.
